Update unit test suite to reflect moved log and config_data objects
The real benefit to this move is the simplification of building and
linking the test suite.  This will almost certainly speed up the
test build, which is always a good thing.
